delete plugin manual registration
