+++ spyder/plugins/switcher/plugin.py
@@ -10,7 +10,7 @@ Switcher Plugin.
 """
 
 # Third-party imports
-from qtpy.QtGui import QIcon
+from qtpy.QtCore import Qt
 
 # Local imports
 from spyder.api.translations import _
@@ -19,16 +19,17 @@ from spyder.api.plugin_registration.decorators import on_plugin_available, on_pl
 from spyder.plugins.switcher.confpage import SwitcherConfigPage
 from spyder.plugins.switcher.container import SwitcherContainer
 from spyder.plugins.mainmenu.api import (ApplicationMenus, FileMenuSections)
-from spyder.utils.qthelpers import (create_action)
 
 
 # --- Constants
 # ----------------------------------------------------------------------------
 class SwitcherActions:
-    FileSwitcherAction = 'file_switcher_action'
-    SymbolFinderAction = 'symbol_finder_action'
+    FileSwitcherAction = 'file switcher'
+    SymbolFinderAction = 'symbol finder'
 
 
+# --- Plugin
+# ----------------------------------------------------------------------------
 class Switcher(SpyderPluginV2):
     """
     Switcher plugin.
@@ -36,7 +37,7 @@ class Switcher(SpyderPluginV2):
 
     NAME = "switcher"
     REQUIRES = []
-    OPTIONAL = [Plugins.MainMenu, Plugins.Shortcuts]
+    OPTIONAL = [Plugins.MainMenu]
     CONTAINER_CLASS = SwitcherContainer
     CONF_WIDGET_CLASS = SwitcherConfigPage
     CONF_SECTION = NAME
@@ -58,7 +59,6 @@ class Switcher(SpyderPluginV2):
 
     def on_initialize(self):
         container = self.get_container()
-        print('Switcher initialized!')
 
          # Switcher shortcuts
         self.create_action(
@@ -66,7 +66,9 @@ class Switcher(SpyderPluginV2):
             _('File switcher...'),
             icon= self.get_icon(),
             tip=_('Fast switch between files'),
-            triggered=self.open_switcher,
+            triggered=container.open_switcher,
+            register_shortcut=True,
+            context=Qt.ApplicationShortcut
         )
         
         self.create_action(
@@ -74,16 +76,18 @@ class Switcher(SpyderPluginV2):
             _('Symbol finder...'),
             icon=self.create_icon('symbol_find'),
             tip=_('Fast symbol search in file'),
-            triggered=self.open_symbolfinder,
+            triggered=container.open_symbolfinder,
+            register_shortcut=True,
+            context=Qt.ApplicationShortcut
         )
     
     @on_plugin_available(plugin=Plugins.MainMenu)
     def on_main_menu_available(self):
         mainmenu = self.get_plugin(Plugins.MainMenu)
         for switcher_action in [SwitcherActions.FileSwitcherAction, SwitcherActions.SymbolFinderAction]:
-            switcher_action = self.get_action(switcher_action)
+            action = self.get_action(switcher_action)
             mainmenu.add_item_to_application_menu(
-                switcher_action,
+                action,
                 menu_id=ApplicationMenus.File,
                 section=FileMenuSections.Switcher,
                 before_section=FileMenuSections.Restart
@@ -93,64 +97,11 @@ class Switcher(SpyderPluginV2):
     def on_main_menu_teardown(self):
         mainmenu = self.get_plugin(Plugins.MainMenu)
         for switcher_action in [SwitcherActions.FileSwitcherAction, SwitcherActions.SymbolFinderAction]:
+            action = self.get_action(switcher_action)
             mainmenu.remove_item_from_application_menu(
-            switcher_action,
+            action,
             menu_id=ApplicationMenus.File)
 
-    @on_plugin_available(plugin=Plugins.Shortcuts)
-    def on_shortcuts_available(self):
-        shortcuts = self.get_plugin(Plugins.Shortcuts)
-        shortcuts.register_shortcut(SwitcherActions.FileSwitcherAction, context="_",
-                            name="File switcher")
-        shortcuts.register_shortcut(SwitcherActions.SymbolFinderAction, context="_",
-                            name="symbol finder", add_shortcut_to_tip=True)
-
-    @on_plugin_teardown(plugin=Plugins.Shortcuts)
-    def on_shortcuts_teardown(self):
-        shortcuts = self.get_plugin(Plugins.Shortcuts)
-        shortcuts.unregister_shortcut(SwitcherActions.FileSwitcherAction, context="_",
-                            name="File switcher")
-        shortcuts.register_shortcut(SwitcherActions.SymbolFinderAction, context="_",
-                            name="symbol finder", add_shortcut_to_tip=True)
-
-
-    def check_compatibility(self):
-        valid = True
-        message = ""  # Note: Remember to use _("") to localize the string
-        return valid, message
-
-    def on_close(self, cancellable=True):
-        return True
-
-
-    def open_switcher(self, symbol=False):
-        """Open switcher dialog box."""
-        switcher = self.get_container().switcher
-        if switcher is not None and switcher.isVisible():
-            switcher.clear()
-            switcher.hide()
-            return
-        if symbol:
-            switcher.set_search_text('@')
-        else:
-            switcher.set_search_text('')
-            switcher.setup()
-        switcher.show()
-
-        # Note: The +6 pixel on the top makes it look better
-        # FIXME: Why is this using the toolbars menu? A: To not be on top of
-        # the toolbars.
-        # Probably toolbars should be taken into account for this 'delta' only
-        # when are visible
-        mainwindow = self.get_main()
-        delta_top = (mainwindow.toolbar.toolbars_menu.geometry().height() +
-                        mainwindow.menuBar().geometry().height() + 6)
-
-        switcher.set_position(delta_top)
-
-    def open_symbolfinder(self):
-        """Open symbol list management dialog box."""
-        self.open_switcher(symbol=True)
 
     # --- Public API
     # ------------------------------------------------------------------------
